Migracion del modal de confirmacion a un componente y adicion de la confirmacion al eliminado de comentarios

+++ app/webroot/js/src/components/Coments/Coment.js
@@ -2,7 +2,7 @@
     const template = /*html*/`
     <b-card :title="coment.user">
         <b-card-text>{{ coment.coment }}</b-card-text>
-        <b-button v-if="showDelete" type="button" variant="danger" @click="deleteComent(coment.id)">Eliminar comentario</b-button>
+        <b-button v-if="showDelete" type="button" variant="danger" @click="deleteComent(coment.id, coment.user)">Eliminar comentario</b-button>
     </b-card>
     `;
 
@@ -10,20 +10,27 @@
 
         methods: {
 
-            deleteComent(id) {
-                fetch(`https://black.digitum.com.mx/retax/blog/practica/coments/${id}`, {
-                    method: 'DELETE'
-                })
-                .then(response => {
+            deleteComent(id, user) {
 
-                    if (response.status === 204) {
-                        this.$emit("deleteComent", id);
-                    }
-                    else {
-                        alert("Error dels ervidor :(");
-                    }
+                store.commit("confirm/showConfirm", {
+                    text: `Estás a punto de eliminar el comentario de <b>${user}</b>. Esta acción no se puede deshacer. ¿Realmente quieres eliminarlo?`,
+                    confirmed: () => {
+                        fetch(`https://black.digitum.com.mx/retax/blog/practica/coments/${id}`, {
+                            method: 'DELETE'
+                        })
+                        .then(response => {
 
+                            if (response.status === 204) {
+                                this.$emit("deleteComent", id);
+                            }
+                            else {
+                                alert("Error dels ervidor :(");
+                            }
+
+                        });
+                    }
                 });
+                
             }
 
         },
